Move more integration models into src/sentry/models/integrations/
