+++ src/sentry/models/integrations/external_issue.py
