+++ src/sentry/models/integrations/integration_feature.py
@@ -7,8 +7,8 @@ from django.utils import timezone
 
 from sentry.db.models import BoundedBigIntegerField, BoundedPositiveIntegerField, Model
 from sentry.db.models.manager import BaseManager
-from sentry.models.integrations import DocIntegration
-from sentry.models.sentryapp import SentryApp
+from sentry.models.integrations.doc_integration import DocIntegration
+from sentry.models.integrations.sentry_app import SentryApp
 
 
 class Feature:
